Version 3.0
Many bugs needed to fix. NEEDS: Allow for all three base robots to be selected. Allow for a counter to keep score of attacks. Allow for player sides of table to toggle correctly. Start up menus give errors explaining there are no values to choose for each map. A stacking annoyance when putting pilot into critical or putting in an expert pilot, it snaps to the card. More clarity around orders of operation needed.